requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
wrapt==1.11.2
